move: move profile tfa setting from UserProfile.tsx to ProfilePage.tsx
