+++ src/pages/profile/ProfilePage.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import { useAuthState } from "../../context/AuthHooks";
 import { UserProps } from "./ProfileProps";
 import UserProfile from "./UserProfile";
-import RecordForm from "./RecordForm";
+import MatchingHistory from "./MatchingHistory";
 import LoadingPage from "../../LoadingPage";
 import UserStats from './UserStats';
 
@@ -30,8 +30,6 @@ const Layout = styled.div`
 const Profile = () => {
   const [userData, setUserData] = useState<UserProps>();
   const { token } = useAuthState();
-  const [tfa, setTfa] = useState(true);
-
   const getData = async() => {
     await axios.get('http://10.12.8.7:3000/user/profile', {
       headers: {
@@ -42,20 +40,6 @@ const Profile = () => {
       setUserData(prev => prev = new UserProps(data.intra_id, data.nickname, data.profile_url, data.two_factor, data.stats, data.matching_history));    
     }).catch(error => {
       alert('user profile loading failed');
-    });
-  }
-
-  const onClickTfa = async () => {
-    await axios.put('http://10.12.8.7:3000/user/profile/tfa', {
-        two_auth: tfa
-    }, {
-          headers: {
-            Authorization:"Bearer " + token
-          }
-    }).then(response => {
-      console.log("set profile image: " + response.status);
-    }).catch(error => {
-      alert('image upload failed');
     });
   }
 
@@ -73,9 +57,8 @@ const Profile = () => {
     <BackGround>
       <Layout>
       <UserProfile data={userData} />
-      <button onClick={() => { onClickTfa() }}>tfa</button>
       { userData.matching_history ?
-        (userData.matching_history.map((value) => <RecordForm
+        (userData.matching_history.map((value) => <MatchingHistory
         key = {value.time}
         time = {value.time}
         result = {value.result}
